Feature/multi tenant (#1)
enable tenant support

+++ mt-access/src/main/java/com/mt/access/application/endpoint/EndpointApplicationService.java
@@ -7,6 +7,7 @@ import com.mt.access.application.endpoint.command.EndpointPatchCommand;
 import com.mt.access.application.endpoint.command.EndpointUpdateCommand;
 import com.mt.access.application.endpoint.representation.EndpointProxyCardRepresentation;
 import com.mt.access.domain.DomainRegistry;
+import com.mt.access.domain.model.AccessDeniedException;
 import com.mt.access.domain.model.cache_profile.CacheProfileId;
 import com.mt.access.domain.model.cache_profile.event.CacheProfileRemoved;
 import com.mt.access.domain.model.cache_profile.event.CacheProfileUpdated;
@@ -20,8 +21,9 @@ import com.mt.access.domain.model.endpoint.Endpoint;
 import com.mt.access.domain.model.endpoint.EndpointId;
 import com.mt.access.domain.model.endpoint.EndpointQuery;
 import com.mt.access.domain.model.endpoint.event.EndpointCollectionModified;
-import com.mt.access.domain.model.system_role.SystemRoleId;
-import com.mt.access.domain.model.system_role.event.SystemRoleDeleted;
+import com.mt.access.domain.model.endpoint.event.SecureEndpointCreated;
+import com.mt.access.domain.model.permission.PermissionId;
+import com.mt.access.domain.model.project.ProjectId;
 import com.mt.common.domain.CommonDomainRegistry;
 import com.mt.common.domain.model.domain_event.AppStarted;
 import com.mt.common.domain.model.domain_event.DomainEventPublisher;
@@ -40,6 +42,8 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import static com.mt.access.domain.model.permission.Permission.*;
+
 @Slf4j
 @Service
 public class EndpointApplicationService {
@@ -56,7 +60,7 @@ public class EndpointApplicationService {
     protected void reloadProxy() {
         if (reloadOnAppStart) {
             try {
-                Thread.sleep(60 * 1000);
+                Thread.sleep(90 * 1000);
             } catch (InterruptedException e) {
                 log.error("wait is interrupted due to", e);
             }
@@ -67,16 +71,27 @@ public class EndpointApplicationService {
 
     @SubscribeForEvent
     @Transactional
-    public String create(EndpointCreateCommand command, String changeId) {
+    public String create(String projectId, EndpointCreateCommand command, String changeId) {
         EndpointId endpointId = new EndpointId();
+        ProjectId projectId1 = new ProjectId(projectId);
+        DomainRegistry.getPermissionCheckService().canAccess(projectId1, CREATE_API);
         return ApplicationServiceRegistry.getApplicationServiceIdempotentWrapper().idempotent(changeId, (change) -> {
             String resourceId = command.getResourceId();
             Optional<Client> client = DomainRegistry.getClientRepository().clientOfId(new ClientId(resourceId));
             if (client.isPresent()) {
                 Client client1 = client.get();
+                if (!client1.getProjectId().equals(projectId1)) {
+                    throw new AccessDeniedException();
+                }
+                boolean secured = command.isSecured();
+                PermissionId permissionId = null;
+                if (secured) {
+                    permissionId = new PermissionId();
+                }
                 Endpoint endpoint = client1.addNewEndpoint(
-                        command.getRoleId() != null ? new SystemRoleId(command.getRoleId()) : null,
+                        permissionId,
                         command.getCacheProfileId() != null ? new CacheProfileId(command.getCacheProfileId()) : null,
+                        command.getName(),
                         command.getDescription(),
                         command.getPath(),
                         endpointId,
@@ -88,6 +103,9 @@ public class EndpointApplicationService {
                 );
                 DomainRegistry.getEndpointRepository().add(endpoint);
                 DomainEventPublisher.instance().publish(new EndpointCollectionModified());
+                if (secured) {
+                    DomainEventPublisher.instance().publish(new SecureEndpointCreated(client1.getProjectId(), endpoint));
+                }
                 return endpointId.getDomainId();
             } else {
                 throw new InvalidClientIdException();
@@ -95,30 +113,43 @@ public class EndpointApplicationService {
         }, ENDPOINT);
     }
 
-    public SumPagedRep<Endpoint> endpoints(String queryParam, String pageParam, String config) {
-        return DomainRegistry.getEndpointRepository().endpointsOfQuery(new EndpointQuery(queryParam, pageParam, config));
+    public SumPagedRep<Endpoint> tenantQuery(String queryParam, String pageParam, String config) {
+        EndpointQuery endpointQuery = new EndpointQuery(queryParam, pageParam, config);
+        DomainRegistry.getPermissionCheckService().canAccess(endpointQuery.getProjectIds(), VIEW_API_SUMMARY);
+        return DomainRegistry.getEndpointRepository().endpointsOfQuery(endpointQuery);
+    }
+    public SumPagedRep<Endpoint> adminQuery(String queryParam, String pageParam, String config) {
+        EndpointQuery endpointQuery = new EndpointQuery(queryParam, pageParam, config);
+        return DomainRegistry.getEndpointRepository().endpointsOfQuery(endpointQuery);
     }
 
-    public Optional<Endpoint> endpoint(String id) {
-        return DomainRegistry.getEndpointRepository().endpointOfId(new EndpointId(id));
+    public Optional<Endpoint> tenantEndpoint(String projectId, String id) {
+        EndpointQuery endpointQuery = new EndpointQuery(new EndpointId(id), new ProjectId(projectId));
+        DomainRegistry.getPermissionCheckService().canAccess(endpointQuery.getProjectIds(), VIEW_API);
+        return DomainRegistry.getEndpointRepository().endpointsOfQuery(endpointQuery).findFirst();
+    }
+    public Optional<Endpoint> adminEndpoint(String id) {
+        EndpointQuery endpointQuery = new EndpointQuery(new EndpointId(id));
+        return DomainRegistry.getEndpointRepository().endpointsOfQuery(endpointQuery).findFirst();
     }
 
     @SubscribeForEvent
     @Transactional
     public void update(String id, EndpointUpdateCommand command, String changeId) {
         log.debug("start of update endpoint");
+        EndpointQuery endpointQuery = new EndpointQuery(new EndpointId(id), new ProjectId(command.getProjectId()));
+        DomainRegistry.getPermissionCheckService().canAccess(endpointQuery.getProjectIds(), EDIT_API);
         EndpointId endpointId = new EndpointId(id);
         ApplicationServiceRegistry.getApplicationServiceIdempotentWrapper().idempotent(changeId, (ignored) -> {
             Optional<Endpoint> endpoint = DomainRegistry.getEndpointRepository().endpointOfId(endpointId);
             if (endpoint.isPresent()) {
                 Endpoint endpoint1 = endpoint.get();
                 endpoint1.update(
-                        command.getRoleId() != null ? new SystemRoleId(command.getRoleId()) : null,
                         command.getCacheProfileId() != null ? new CacheProfileId(command.getCacheProfileId()) : null,
+                        command.getName(),
                         command.getDescription(),
                         command.getPath(),
                         command.getMethod(),
-                        command.isSecured(),
                         command.isWebsocket(),
                         command.isCsrfEnabled(),
                         command.getCorsProfileId() != null ? new CORSProfileId(command.getCorsProfileId()) : null
@@ -133,10 +164,11 @@ public class EndpointApplicationService {
 
     @SubscribeForEvent
     @Transactional
-    public void removeEndpoint(String id, String changeId) {
-        EndpointId endpointId = new EndpointId(id);
+    public void removeEndpoint(String projectId, String id, String changeId) {
+        EndpointQuery endpointQuery = new EndpointQuery(new EndpointId(id), new ProjectId(projectId));
+        DomainRegistry.getPermissionCheckService().canAccess(endpointQuery.getProjectIds(), DELETE_API);
         ApplicationServiceRegistry.getApplicationServiceIdempotentWrapper().idempotent(changeId, (ignored) -> {
-            Optional<Endpoint> endpoint = DomainRegistry.getEndpointRepository().endpointOfId(endpointId);
+            Optional<Endpoint> endpoint = DomainRegistry.getEndpointRepository().endpointsOfQuery(endpointQuery).findFirst();
             if (endpoint.isPresent()) {
                 Endpoint endpoint1 = endpoint.get();
                 DomainRegistry.getEndpointRepository().remove(endpoint1);
@@ -148,9 +180,19 @@ public class EndpointApplicationService {
 
     @SubscribeForEvent
     @Transactional
-    public void removeEndpoints(String queryParam, String changeId) {
+    public void removeEndpoints(String projectId, String queryParam, String changeId) {
+        ProjectId projectId1 = new ProjectId(projectId);
+        DomainRegistry.getPermissionCheckService().canAccess(projectId1, BATCH_DELETE_API);
         ApplicationServiceRegistry.getApplicationServiceIdempotentWrapper().idempotent(changeId, (change) -> {
-            Set<Endpoint> allByQuery = QueryUtility.getAllByQuery((query) -> DomainRegistry.getEndpointRepository().endpointsOfQuery((EndpointQuery) query), new EndpointQuery(queryParam));
+            Set<Endpoint> allByQuery = QueryUtility.getAllByQuery((query) -> DomainRegistry.getEndpointRepository().endpointsOfQuery((EndpointQuery) query), new EndpointQuery(queryParam, projectId1));
+            Set<ProjectId> collect = allByQuery.stream().map(Endpoint::getProjectId).collect(Collectors.toSet());
+            if (collect.size() != 1) {
+                throw new AccessDeniedException();
+            }
+            ProjectId[] a = new ProjectId[1];
+            if (!collect.toArray(a)[0].equals(projectId1)) {
+                throw new AccessDeniedException();
+            }
             DomainRegistry.getEndpointRepository().remove(allByQuery);
             DomainEventPublisher.instance().publish(
                     new EndpointCollectionModified()
@@ -161,21 +203,22 @@ public class EndpointApplicationService {
 
     @SubscribeForEvent
     @Transactional
-    public void patchEndpoint(String id, JsonPatch command, String changeId) {
+    public void patchEndpoint(String projectId, String id, JsonPatch command, String changeId) {
+        ProjectId projectId1 = new ProjectId(projectId);
+        DomainRegistry.getPermissionCheckService().canAccess(projectId1, PATCH_API);
         EndpointId endpointId = new EndpointId(id);
         ApplicationServiceRegistry.getApplicationServiceIdempotentWrapper().idempotent(changeId, (ignored) -> {
-            Optional<Endpoint> endpoint = DomainRegistry.getEndpointRepository().endpointOfId(endpointId);
+            Optional<Endpoint> endpoint = DomainRegistry.getEndpointRepository().endpointsOfQuery(new EndpointQuery(endpointId, projectId1)).findFirst();
             if (endpoint.isPresent()) {
                 Endpoint endpoint1 = endpoint.get();
                 EndpointPatchCommand beforePatch = new EndpointPatchCommand(endpoint1);
                 EndpointPatchCommand afterPatch = CommonDomainRegistry.getCustomObjectSerializer().applyJsonPatch(command, beforePatch, EndpointPatchCommand.class);
                 endpoint1.update(
-                        endpoint1.getSystemRoleId(),
                         endpoint1.getCacheProfileId(),
+                        afterPatch.getName(),
                         afterPatch.getDescription(),
                         afterPatch.getPath(),
                         afterPatch.getMethod(),
-                        endpoint1.isSecured(),
                         endpoint1.isWebsocket(),
                         endpoint1.isCsrfEnabled(),
                         endpoint1.getCorsProfileId()
@@ -242,27 +285,6 @@ public class EndpointApplicationService {
         }, ENDPOINT);
     }
 
-    /**
-     * refresh proxy when referred role group is updated
-     *
-     * @param deserialize
-     */
-    @SubscribeForEvent
-    @Transactional
-    public void handle(SystemRoleDeleted deserialize) {
-        ApplicationServiceRegistry.getApplicationServiceIdempotentWrapper().idempotent(deserialize.getId().toString(), (ignored) -> {
-            log.debug("handle role group removed");
-            SystemRoleId systemRoleId = new SystemRoleId(deserialize.getDomainId().getDomainId());
-            Set<Endpoint> allByQuery = QueryUtility.getAllByQuery((query) -> DomainRegistry.getEndpointRepository().endpointsOfQuery((EndpointQuery) query), new EndpointQuery(systemRoleId));
-            if (!allByQuery.isEmpty()) {
-                DomainEventPublisher.instance().publish(new EndpointCollectionModified());
-                allByQuery.forEach(e -> e.setSystemRoleId(null));
-            }
-            return null;
-        }, ENDPOINT);
-
-    }
-
     @SubscribeForEvent
     @Transactional
     public void handle(CacheProfileRemoved deserialize) {
@@ -294,8 +316,8 @@ public class EndpointApplicationService {
         }, ENDPOINT);
     }
 
-    public SumPagedRep<EndpointProxyCardRepresentation> proxyEndpoints(String queryParam, String pageParam, String config) {
-        SumPagedRep<Endpoint> endpoints = endpoints(queryParam, pageParam, config);
+    public SumPagedRep<EndpointProxyCardRepresentation> internalQuery(String queryParam, String pageParam, String config) {
+        SumPagedRep<Endpoint> endpoints = DomainRegistry.getEndpointRepository().endpointsOfQuery(new EndpointQuery(queryParam, pageParam, config));
         List<EndpointProxyCardRepresentation> collect = endpoints.getData().stream().map(EndpointProxyCardRepresentation::new).collect(Collectors.toList());
         EndpointProxyCardRepresentation.updateDetail(collect);
         return new SumPagedRep<>(collect, endpoints.getTotalItemCount());
